+++ apps/api/src/app/order/interfaces/activities.interface.ts
@@ -9,7 +9,7 @@ export interface Activities {
 }
 
 export interface Activity extends Order {
-  Account?: AccountWithPlatform;
+  account?: AccountWithPlatform;
   error?: ActivityError;
   feeInAssetProfileCurrency: number;
   feeInBaseCurrency: number;