+++ apps/api/src/app/order/order.service.ts
@@ -100,10 +100,10 @@ export class OrderService {
       userId: string;
     }
   ): Promise<Order> {
-    let Account: Prisma.AccountCreateNestedOneWithoutActivitiesInput;
+    let account: Prisma.AccountCreateNestedOneWithoutActivitiesInput;
 
     if (data.accountId) {
-      Account = {
+      account = {
         connect: {
           id_userId: {
             userId: data.userId,
@@ -179,7 +179,7 @@ export class OrderService {
     const order = await this.prismaService.order.create({
       data: {
         ...orderData,
-        Account,
+        account,
         isDraft,
         tags: {
           connect: tags.map(({ id }) => {
@@ -475,8 +475,8 @@ export class OrderService {
 
     if (withExcludedAccounts === false) {
       where.OR = [
-        { Account: null },
-        { Account: { NOT: { isExcluded: true } } }
+        { account: null },
+        { account: { NOT: { isExcluded: true } } }
       ];
     }
 
@@ -487,8 +487,7 @@ export class OrderService {
         take,
         where,
         include: {
-          // eslint-disable-next-line @typescript-eslint/naming-convention
-          Account: {
+          account: {
             include: {
               platform: true
             }
@@ -650,8 +649,8 @@ export class OrderService {
     if (['FEE', 'INTEREST', 'ITEM', 'LIABILITY'].includes(data.type)) {
       delete data.SymbolProfile.connect;
 
-      if (data.Account?.connect?.id_userId?.id === null) {
-        data.Account = { disconnect: true };
+      if (data.account?.connect?.id_userId?.id === null) {
+        data.account = { disconnect: true };
       }
     } else {
       delete data.SymbolProfile.update;